chore: Change widget order in ContainerWidget
Previously both the order in which actions like clicks were processed as
well as the order in which widgets were rendered would be in insertion
order.

This is problematic, as the widgets that are on top (drawn last) are not
the ones receiving input first, so you might be clicking something
entirely different than you might expect.

It is now changed so that the widgets are still rendered in insertion
order, but input is processed in reverse insertion order.