chore: Workaround for build issues
Once they language files get above a certain size (which they have), the
builds stall.

The reason for this is that the way I've set it up right now, it first
reads the entire source file before starting to process the data, which
means the buffer needs to be large enough to contain the entire file.

For now, I've just increased the buffer size, which should avoid this
for the time being.