chore: Fixed relative position inside a container
The positioning functions don't have to take the position of the
container they are in into account, this is added later.

However, when getting the position of the widget the positioning should
be relative to, the position of the container it is in IS included
already, causing the parent position to be added to the relative
widget's position twice.